+++ lib/ret.ex
@@ -1,9 +1,19 @@
 defmodule Ret do
-  @moduledoc """
-  Ret keeps the contexts that define your domain
-  and business logic.
+  import Canada, only: [can?: 2]
+  alias Ret.{Account, Repo}
 
-  Contexts are also responsible for managing your data, regardless
-  if it comes from the database, an external API or others.
-  """
+  def get_account_by_id(account_id) do
+    Repo.get(Account, account_id)
+  end
+
+  def delete_account(%Account{} = acting_account, %Account{} = account_to_delete) do
+    if can?(acting_account, delete_account(account_to_delete)) do
+      case Repo.delete(account_to_delete) do
+        {:ok, _} -> :ok
+        {:error, _} -> {:error, :failed}
+      end
+    else
+      {:error, :forbidden}
+    end
+  end
 end
